# 8276220: Reduce excessive allocations in DateTimeFormatter