--- a/test/jdk/java/time/test/java/time/format/TestFractionPrinterParser.java
+++ b/test/jdk/java/time/test/java/time/format/TestFractionPrinterParser.java
@@ -69,7 +69,6 @@ import java.text.ParsePosition;
 import java.time.DateTimeException;
 import java.time.LocalTime;
 import java.time.format.DateTimeFormatter;
-import java.time.format.DateTimeFormatterBuilder;
 import java.time.temporal.TemporalAccessor;
 import java.time.temporal.TemporalField;