--- a/test/micro/org/openjdk/bench/java/time/format/DateTimeFormatterBench.java
+++ b/test/micro/org/openjdk/bench/java/time/format/DateTimeFormatterBench.java
@@ -25,9 +25,6 @@ import org.openjdk.jmh.annotations.Setup;
 import org.openjdk.jmh.annotations.State;
 import org.openjdk.jmh.annotations.Warmup;
 import org.openjdk.jmh.infra.Blackhole;
-import org.openjdk.jmh.runner.Runner;
-import org.openjdk.jmh.runner.options.Options;
-import org.openjdk.jmh.runner.options.OptionsBuilder;
 
 @BenchmarkMode(Mode.Throughput)
 @OutputTimeUnit(TimeUnit.MILLISECONDS)
@@ -99,5 +96,4 @@ public class DateTimeFormatterBench {
             blackhole.consume(stringBuilder);
         }
     }
-
 }