Prompted by a request from Volkan Yazıcı I took a look at why the java.time formatters are less efficient for some common patterns than custom formatters in apache-commons and log4j. This patch reduces the gap, without having looked at the third party implementations. 

When printing times:
- Avoid turning integral values into `String`s before appending them to the buffer 
- Specialize `appendFraction` for `NANO_OF_SECOND` to avoid use of `BigDecimal`

This means a speed-up and reduction in allocations when formatting almost any date or time pattern, and especially so when including sub-second parts (`S-SSSSSSSSS`).

Much of the remaining overhead can be traced to the need to create a `DateTimePrintContext` and adjusting `Instant`s into a `ZonedDateTime` internally. We could likely also win performance by specializing some common patterns.

Testing: tier1-3

<!-- Anything below this marker will be automatically updated, please do not edit manually! -->
---------
### Progress
- [x] Change must not contain extraneous whitespace
- [x] Commit message must refer to an issue
- [x] Change must be properly reviewed

### Issue
 * [JDK-8276220](https://bugs.openjdk.java.net/browse/JDK-8276220): Reduce excessive allocations in DateTimeFormatter


### Reviewers
 * [Stephen Colebourne](https://openjdk.java.net/census#scolebourne) (@jodastephen - Author) ⚠️ Review applies to 1d21a1a50e4724eb9b566d8c3142cc6fd3145d75
 * [Naoto Sato](https://openjdk.java.net/census#naoto) (@naotoj - **Reviewer**)


### Reviewing
<details><summary>Using <code>git</code></summary>

Checkout this PR locally: \
`$ git fetch https://git.openjdk.java.net/jdk pull/6188/head:pull/6188` \
`$ git checkout pull/6188`

Update a local copy of the PR: \
`$ git checkout pull/6188` \
`$ git pull https://git.openjdk.java.net/jdk pull/6188/head`

</details>
<details><summary>Using Skara CLI tools</summary>

Checkout this PR locally: \
`$ git pr checkout 6188`

View PR using the GUI difftool: \
`$ git pr show -t 6188`

</details>
<details><summary>Using diff file</summary>

Download this PR as a diff file: \
<a href="https://git.openjdk.java.net/jdk/pull/6188.diff">https://git.openjdk.java.net/jdk/pull/6188.diff</a>

</details>