@cl4es 2021-11-01
Microbenchmark numbers for the supplied `DateTimeFormatterBench`. 

Baseline:
```
Benchmark                                                   (pattern)   Mode  Cnt  Score   Error   Units
DateTimeFormatterBench.formatInstants                        HH:mm:ss  thrpt   10  5.904 ± 0.221  ops/ms
DateTimeFormatterBench.formatInstants                    HH:mm:ss.SSS  thrpt   10  2.654 ± 0.096  ops/ms
DateTimeFormatterBench.formatInstants           yyyy-MM-dd'T'HH:mm:ss  thrpt   10  3.463 ± 0.302  ops/ms
DateTimeFormatterBench.formatInstants       yyyy-MM-dd'T'HH:mm:ss.SSS  thrpt   10  2.002 ± 0.120  ops/ms
DateTimeFormatterBench.formatZonedDateTime                   HH:mm:ss  thrpt   10  7.899 ± 0.270  ops/ms
DateTimeFormatterBench.formatZonedDateTime               HH:mm:ss.SSS  thrpt   10  2.806 ± 0.223  ops/ms
DateTimeFormatterBench.formatZonedDateTime      yyyy-MM-dd'T'HH:mm:ss  thrpt   10  4.031 ± 0.377  ops/ms
DateTimeFormatterBench.formatZonedDateTime  yyyy-MM-dd'T'HH:mm:ss.SSS  thrpt   10  2.027 ± 0.144  ops/ms
```
Patched:
```
Benchmark                                                   (pattern)   Mode  Cnt  Score   Error   Units
DateTimeFormatterBench.formatInstants                        HH:mm:ss  thrpt   10  6.805 ± 0.356  ops/ms
DateTimeFormatterBench.formatInstants                    HH:mm:ss.SSS  thrpt   10  5.091 ± 0.391  ops/ms
DateTimeFormatterBench.formatInstants           yyyy-MM-dd'T'HH:mm:ss  thrpt   10  4.487 ± 0.232  ops/ms
DateTimeFormatterBench.formatInstants       yyyy-MM-dd'T'HH:mm:ss.SSS  thrpt   10  3.301 ± 0.160  ops/ms
DateTimeFormatterBench.formatZonedDateTime                   HH:mm:ss  thrpt   10  9.531 ± 0.752  ops/ms
DateTimeFormatterBench.formatZonedDateTime               HH:mm:ss.SSS  thrpt   10  6.329 ± 0.260  ops/ms
DateTimeFormatterBench.formatZonedDateTime      yyyy-MM-dd'T'HH:mm:ss  thrpt   10  5.356 ± 0.087  ops/ms
DateTimeFormatterBench.formatZonedDateTime  yyyy-MM-dd'T'HH:mm:ss.SSS  thrpt   10  4.037 ± 0.126  ops/ms
```