@cl4es 2021-11-03
Thanks for reviewing, @jodastephen! 

I think a spec update sounds good, but I think that should be done in as a follow-up. If you would be willing to provide the wording for such a spec update I can take care of creating a CSR.