@jodastephen 2021-11-09
@cl4es For `DateTimeFormatterBuilder ` spec, I propose changing the existing wording slightly

    If the field value in the date-time to be printed is invalid it
    cannot be printed and an exception will be thrown.

to 

    If the field value in the date-time to be printed is outside the
    range of valid values then an exception will be thrown.